function that returns the natural square root of a number

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -26,7 +26,7 @@ int _sqrt_recursion(int n)
  */
 int find_sqrt(int start, int end, int n)
 {
-	int mid;
+	long mid;
 
 	if (end >= start)
 	{
